Create hasMutation function to decide if a DNA secuence has mutations
This function runs through the entire vector and verify each one of the
characters.
First verify if it's a right character using isValid function.
Then check if the character is included into mutations array.
And finally verify each one of the validations to find if this character
is part of a mutation.
When a mutation is founded the counter ir increased.
At the end of the program, if the counter is greater than or equal to 2
return true because is a mutation, if it doesn't return false.